pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "tiktok-geo-regulation"
version = "1.0.0"
description = "TikTok Geo-Regulation AI System - Automated compliance analysis for TikTok features across global jurisdictions"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "TikTok TechJam 2025 Team" },
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
]
dependencies = [
    # Core Framework
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.20.0",
    "pydantic>=2.0.0",
    "python-multipart>=0.0.6",
    # LangChain & LangGraph - latest compatible versions
    "langgraph>=0.2.0",
    "langchain>=0.3.0",
    "langchain-anthropic>=0.2.0",
    "langchain-openai>=0.2.0",
    "langchain-google-genai>=2.0.0",
    # Google Gemini
    "google-generativeai>=0.4.1",
    # Database
    "sqlalchemy>=2.0.0",
    "alembic>=1.12.0",
    "asyncpg>=0.29.0",
    "psycopg2-binary>=2.9.0",
    # Caching
    "redis>=5.0.0",
    "aioredis>=2.0.0",
    # HTTP Client
    "aiohttp>=3.9.0",
    "httpx>=0.25.0",
    "aiofiles>=23.0.0",
    # Frontend
    "streamlit>=1.28.0",
    "plotly>=5.17.0",
    "pandas>=2.1.0",
    # Utilities
    "python-dotenv>=1.0.0",
    "pydantic-settings>=2.1.0",
    "typing-extensions>=4.11.0",
    # Security
    "cryptography>=42.0.0",
    "python-jose[cryptography]>=3.3.0",
    # Monitoring
    "prometheus-client>=0.19.0",
    "streamlit-mermaid>=0.3.0",
    # MCP Support
    "mcp>=1.0.0",
    "requests>=2.31.0",
    # Real MCP Dependencies
    "chromadb>=0.4.0",
    "sentence-transformers>=2.2.0",
    "pypdf2>=3.0.0",
    # Legal MCP Additional Dependencies
    "PyMuPDF>=1.23.0",
    "pgvector>=0.2.0",
    "numpy>=1.24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.11.0",
    "isort>=5.12.0",
    "mypy>=1.7.0",
]

[project.scripts]
geolegal-api = "src.main:main"
geolegal-ui = "src.ui.app:main"

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 88

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
